#6 Feat/Integration-page-connexion: externalisation fonctions JS. ajout import/export et fichiers.
