+++ FrontEnd/js/hasValidateToken.js
@@ -1,0 +1,4 @@
+
+const hasValidateToken = (data) => data?.userId ? true : false;
+
+export {hasValidateToken}
