+++ FrontEnd/pages/login.html
@@ -67,6 +67,6 @@
         </ul>
       </nav>
     </footer>
-    <script src="../js/LogUser.js"></script>
+    <script type="module" src="../js/LogUser.js"></script>
   </body>
 </html>
